feat: Create WebConfig and change type on buildingArea int to floor
cors 관련 Config 생성, buildingArea 소수형으로 형변환

+++ src/main/java/com/project/dabang/domain/construction/Construction.java
@@ -25,7 +25,7 @@ public class Construction extends Timestamped {
     @Enumerated(EnumType.STRING)
     private RoomType roomType;
 
-    private int buildingArea;
+    private float buildingArea;
 
     private int floor;
